% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGA-Optimizer.R
\name{SCGA}
\alias{SCGA}
\title{Run the SCGA optimisation}
\usage{
SCGA(control = list(), ...)
}
\arguments{
\item{control}{list of controls. See \code{\link{Initialise}}}
}
\value{
list
}
\description{
Structured-Chromosome Genetic Algorithm
}
\examples{

}
