% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGA-Initialise.R
\name{Initialise}
\alias{Initialise}
\title{This function returns the default controls and other useful variables.
 Control is a list of the settings:}
\usage{
Initialise(control = list(), ...)
}
\arguments{
\item{convergence}{Stopping criterion: absolute difference between the current best and
the known minimum}

\item{cpus}{numeric. indicatig number of cores over which parallelise}

\item{creatCandFun}{function. See \code{\link{createCandidate}}}

\item{createMutFun}{function. See \code{\link{createMutFun}}}

\item{crossFun}{function. See \code{\link{crossFun}}}

\item{dontChangeCross}{numeric vector. Feature number that not undergo to Crossover}

\item{dontChangeMut}{numeric vector. Feature number that not undergo to Mutation}

\item{elitism}{numeric. Number of candidates to preserve to the next population. Default is size / 10}

\item{evaluatePopDF}{function. See \code{\link{evaluatePopDF}}}

\item{feature.}{list or function that creates the list. See \code{\link{feature}}}

\item{fitnessFN}{function. Receives the observations of the objective functions and returns}

\item{Fun}{function. Objective function
a vector of the same length repesententing the fitness. Default is Ranking fitness.}
}
\description{
This function returns the default controls and other useful variables.
 Control is a list of the settings:
}
